refactored leads app
