+++ leads/templates/leads/landing-page.html
@@ -76,8 +76,8 @@
 {% comment %} TEAM SECTION {% endcomment %}
 
 <section class="text-gray-600 body-font ">
-  <div class="container px-5 py-20 mx-auto">
-    <div class="flex flex-col text-center w-full mb-20">
+  <div class="container px-5 py-17 mx-auto mb-10">
+    <div class="flex flex-col text-center w-full ">
       <h1
         class="mb-4 text-4xl font-extrabold tracking-tight leading-none text-gray-900 md:text-5xl lg:text-6xl dark:text-white"
       >
