+++ leads/templates/leads/leads/lead-list.html
@@ -1,8 +1,5 @@
 {% extends 'base.html' %} {% load static %} {% block content %}
 
-{% comment %} https://flowbite.com/docs/components/tables/  {% endcomment %}
-
-
 <main>
     <section class="text-gray-600 body-font overflow-hidden mt-0">
       <div>{% include 'snippets/alert.html' %}</div>
