--- songs/templates/songs/footer.html
@@ -1,46 +1,0 @@
-{% load static %}
-
-<footer>
-    <div class="container">
-        <div class="row row-bottom-paddiing-md">
-            <div class="col-md-3">
-                <h3 class="footer-title">All About Music</h3>
-                
-            </div>
-
-        </div>
-        <div class="row">
-            <div class="col-md-12">
-                <p class="fh5co-social-icons">
-                    <a href="#" target="_blank"><i class="icon-twitter2"></i></a>
-                    <a href="#" target="_blank"><i class="icon-facebook2"></i></a>
-                    <a href="#" target="_blank"><i class="icon-instagram"></i></a>
-                    <a href="#" target="_blank"><i class="icon-linkedin"></i></a>
-                   
-                </p>
-                <div class="container-fluid copyright-section">
-                    <div class="container">
-                      <div class="row">
-                        <div class="col-md-6 text-center mb-3 text-md-start mb-md-0">
-                          &copy; <a href="#">Oriri Restaurant</a>, All Right Reserved,
-                          <script>
-                            document.write(new Date().getFullYear());
-                          </script>
-                          .
-                        </div>
-                        <div class="col-md-6 text-md-end text-center">
-                          Developed By
-                          <a href="https:/linkedin/blaise-ejikeme" target="_blank"
-                            >Blaisemart</a
-                          >
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-           
-            </div>
-        </div>
-    </div>
-
-
-</footer>
