+++ songs/templates/songs/navigation.html
@@ -1,7 +1,7 @@
 {% load static %}
 <header class="max-w-screen-xl px-3 lg:py-3 z-10 relative">
   <nav class="bg-white dark:bg-blue-900 font-medium border-gray-200 dark:bg-gray-900 dark:border-gray-700">
-    <div class="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto p-4">
+    <div class="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto pt-4 px-4">
       <a class="flex items-center space-x-3 rtl:space-x-reverse">
           <img src="{% static "images/logo.jpg" %}" class="h-10" alt="medconnect logo" />
           <span class="self-center text-2xl font-semibold whitespace-nowrap dark:text-white">MED-CONNECT™</span>
