+++ templates/account/email.html
@@ -1,5 +1,5 @@
-{% extends "account/base.html" %} {% load i18n %} {% block head_title %}{%
-trans"E-mail Addresses" %} {% endblock %} {% block content %}
+{% extends "account/base.html" %} {% load i18n %} {% block head_title %}
+{%trans"E-mail Addresses" %} {% endblock %} {% block content %}
 <main class="max-w-lg mx-auto">
   <div class="py-5 border-b border-gray-200">
     <section class="border-b border-gray-200">
@@ -20,8 +20,7 @@ trans"E-mail Addresses" %} {% endblock %} {% block content %}
               class="{% if emailaddress.primary %}primary_email {%endif%}"
             >
               <input id="email_radio_{{forloop.counter}}" type="radio"
-              name="email" {% if emailaddress.primary or
-              user.emailaddress_set.count == 1 %}checked="checked"{%endif %}
+              name="email" {% if emailaddress.primary or user.emailaddress_set.count == 1 %}checked="checked"{%endif %}
               value="{{emailaddress.email}}"/> {{ emailaddress.email }} {% if
               emailaddress.verified %}
               <span class="verified">{% trans "Verified" %}</span>
@@ -88,8 +87,7 @@ trans"E-mail Addresses" %} {% endblock %} {% block content %}
 {% endblock %} {% block extra_body %}
 <script type="text/javascript">
   (function () {
-    var message =
-      "{% trans 'Do you really want to remove the selected e-mail address?' %}";
+    var message = "{% trans 'Do you really want to remove the selected e-mail address?' %}";
     var actions = document.getElementsByName("action_remove");
     if (actions.length) {
       actions[0].addEventListener("click", function (e) {
