+++ templates/base.html
@@ -26,7 +26,8 @@
     tailwind_css %} {% endcomment %} {% block extra_css%} {% endblock extra_css%}
     <title>Lead Management System</title>
     <link
-      href="https://fonts.googleapis.com/css?family=Material+Icons|Material+Icons+Outlined|Material+Icons+Two+Tone|Material+Icons+Round|Material+Icons+Sharp"
+      href="https://fonts.googleapis.com/css?family=Material+Icons|Material+Icons+Outlined|\
+      Material+Icons+Two+Tone|Material+Icons+Round|Material+Icons+Sharp"
       rel="stylesheet"
     />
     <link
@@ -54,9 +55,5 @@
     {% block extra_scripts %} {% endblock extra_scripts %} {% include 'scripts.html' %}
     <script src="{% static  'js/main.js' %}"></script>
     <script src="{% static  'js/chat.js' %}"></script>
-     <section class="">
-      <div class="max-w-screen-xl lg:py-3 z-10 relative">
-      
-     
   </body>
 </html>
