+++ templates/footer.html
@@ -3,7 +3,6 @@
     <footer class="text-gray-700 body-font bg-gray-500 font-medium">
       <div class="mx-auto w-full max-w-screen-xl p-4 py-8 lg:py-8">
         <div class="md:flex md:justify-between">
-          <div class="mb-6 md:mb-0"></div>
           <div class="grid grid-cols-2 gap-8 sm:gap-6 sm:grid-cols-3">
             <div>
               <h2 class="mb-6 text-sm font-semibold text-gray-900 uppercase">
